refactor(users): refine profile editing UI
